refactor(wallet): Share VariablesProvider, injecting persistence
Second slice of #894, and the one that tests whether the contexts are
shareable at all -- everything in the dependent tiers waits on that answer.

The two copies differed by 107 lines, and the difference was one idea: the
extension's popup is destroyed every time it closes, so it writes each value
through to chrome storage and reads it back on open, while react-wallet keeps a
live page and persists nothing. That made five of its setters async and added
storeState, resetCredentialState, refreshCredentialsStored and
refreshRegistryStored, plus heldDID and manifest, to its shape.

So persistence is injected, like the snackbar's safe-area inset: the shared
provider takes an optional `store`, and every setter writes through it when
present. The extension passes a function that messages its background script --
and passes nothing for the full-page view, which is what its `isBrowser` check
was expressing inside the provider. react-wallet omits it and keeps exactly the
behaviour it had.

Answering the question the slice was for: this context was shareable without
compromise, and the divergence turned out to be a capability rather than a
disagreement about state. That is encouraging for UIContext and WalletProvider,
though both are larger and neither is guaranteed to be the same shape.

react-wallet needed no call-site changes. Its persisted setters go from
Dispatch<SetStateAction<string>> to (value) => Promise<void>, which is safe
because nothing in either app calls them with the updater form.

44 files repointed. Same-named pairs: 38 -> 37, diverged lines 5649 -> 5547.

Refs #894
