+++ apps/browser-extension/src/components/CloneAssetTab.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { Box, TextField, Select, MenuItem, Button } from "@mui/material";
 import { useWalletContext } from "../contexts/WalletProvider";
-import { useVariablesContext } from "../contexts/VariablesProvider";
+import { useVariablesContext } from "@didcid/wallet-ui";
 import { useUIContext } from "../contexts/UIContext";
 import { useSnackbar } from "@didcid/wallet-ui";
