+++ apps/browser-extension/src/components/HeldTab.tsx
@@ -3,7 +3,7 @@ import { Box, Button, TextField } from "@mui/material";
 import WarningModal from "../modals/WarningModal";
 import { useWalletContext } from "../contexts/WalletProvider";
 import { useSnackbar } from "@didcid/wallet-ui";
-import { useVariablesContext } from "../contexts/VariablesProvider";
+import { useVariablesContext } from "@didcid/wallet-ui";
 import { useUIContext } from "../contexts/UIContext";
 import JsonViewer from "./JsonViewer";
 import { requestBrowserRefresh } from "../utils/utils";
