+++ apps/browser-extension/src/components/NostrApproval.tsx
@@ -1,7 +1,7 @@
 import { useCallback, useEffect, useRef, useState } from "react";
 import { Box, Button, Checkbox, FormControlLabel, Typography, CircularProgress } from "@mui/material";
 import { useWalletContext } from "../contexts/WalletProvider";
-import { useVariablesContext } from "../contexts/VariablesProvider";
+import { useVariablesContext } from "@didcid/wallet-ui";
 
 interface NostrApprovalProps {
     requestId: string;
