+++ apps/browser-extension/src/components/SchemaTab.tsx
@@ -2,7 +2,7 @@ import { useState } from "react";
 import { Box, Button, IconButton, MenuItem, Select, TextField, Tooltip } from "@mui/material";
 import { useWalletContext } from "../contexts/WalletProvider";
 import { useUIContext } from "../contexts/UIContext";
-import { useVariablesContext } from "../contexts/VariablesProvider";
+import { useVariablesContext } from "@didcid/wallet-ui";
 import { Edit } from "@mui/icons-material";
 import TextInputModal from "../modals/TextInputModal";
 import CopyResolveDID from "./CopyResolveDID";
