+++ apps/browser-extension/src/components/VaultTab.tsx
@@ -14,7 +14,7 @@ import {
     Edit,
 } from "@mui/icons-material";
 import { useWalletContext } from "../contexts/WalletProvider";
-import { useVariablesContext } from "../contexts/VariablesProvider";
+import { useVariablesContext } from "@didcid/wallet-ui";
 import { useUIContext } from "../contexts/UIContext";
 import { useSnackbar } from "@didcid/wallet-ui";
 import AuthDetailsModal from "../modals/AuthDetailsModal";
