+++ apps/react-wallet/src/components/FileTab.tsx
@@ -3,7 +3,7 @@ import { Box, Button, FormControl, IconButton, LinearProgress, MenuItem, Select,
 import { Download, Edit } from "@mui/icons-material";
 import { useWalletContext } from "../contexts/WalletProvider";
 import { useUIContext } from "../contexts/UIContext";
-import { useVariablesContext } from "../contexts/VariablesProvider";
+import { useVariablesContext } from "@didcid/wallet-ui";
 import { useSnackbar } from "@didcid/wallet-ui";
 import { FileAsset } from "@didcid/keymaster/types";
 import { DidCidDocument } from "@didcid/gatekeeper/types";
