+++ apps/react-wallet/src/components/GroupsTab.tsx
@@ -12,7 +12,7 @@ import {
     Typography
 } from "@mui/material";
 import { useWalletContext } from "../contexts/WalletProvider";
-import { useVariablesContext } from "../contexts/VariablesProvider";
+import { useVariablesContext } from "@didcid/wallet-ui";
 import { useUIContext } from "../contexts/UIContext";
 import { useSnackbar } from "@didcid/wallet-ui";
 import {Delete, Edit} from "@mui/icons-material";
