+++ apps/react-wallet/src/components/IdentitiesTab.tsx
@@ -14,7 +14,7 @@ import { useSnackbar } from "@didcid/wallet-ui";
 import WarningModal from "../modals/WarningModal";
 import TextInputModal from "../modals/TextInputModal";
 import { SelectInputModal } from "@didcid/wallet-ui";
-import { useVariablesContext } from "../contexts/VariablesProvider";
+import { useVariablesContext } from "@didcid/wallet-ui";
 import type { AddressCheckResult, AddressInfo, FileAsset, ImageAsset, NostrKeys } from "@didcid/keymaster/types";
 import GatekeeperClient from "@didcid/clients/gatekeeper";
 import {
