+++ apps/react-wallet/src/components/IssueTab.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import { useWalletContext } from "../contexts/WalletProvider";
-import { useVariablesContext } from "../contexts/VariablesProvider";
+import { useVariablesContext } from "@didcid/wallet-ui";
 import { useSnackbar } from "@didcid/wallet-ui";
 import { CredentialForm } from "@didcid/wallet-ui";
 import {
