+++ apps/react-wallet/src/components/PollTab.tsx
@@ -28,7 +28,7 @@ import {
     PersonAdd,
 } from "@mui/icons-material";
 import { useWalletContext } from "../contexts/WalletProvider";
-import { useVariablesContext } from "../contexts/VariablesProvider";
+import { useVariablesContext } from "@didcid/wallet-ui";
 import { useUIContext } from "../contexts/UIContext";
 import { useSnackbar } from "@didcid/wallet-ui";
 import PollResultsModal from "../modals/PollResultsModal";
