+++ apps/react-wallet/src/components/VaultTab.tsx
@@ -15,7 +15,7 @@ import {
     Close,
 } from "@mui/icons-material";
 import { useWalletContext } from "../contexts/WalletProvider";
-import { useVariablesContext } from "../contexts/VariablesProvider";
+import { useVariablesContext } from "@didcid/wallet-ui";
 import { useUIContext } from "../contexts/UIContext";
 import { useSnackbar } from "@didcid/wallet-ui";
 import AuthDetailsModal from "../modals/AuthDetailsModal";
