+++ apps/react-wallet/src/contexts/UIContext.tsx
@@ -17,7 +17,7 @@ import {
     Token,
 } from '@mui/icons-material';
 import { useWalletContext } from "./WalletProvider";
-import { useVariablesContext } from "./VariablesProvider";
+import { useVariablesContext } from "@didcid/wallet-ui";
 import { useSnackbar } from "@didcid/wallet-ui";
 import WalletWeb from "@didcid/keymaster/wallet/web";
