+++ packages/wallet-ui/src/index.ts
@@ -2,3 +2,5 @@ export { SnackbarProvider, useSnackbar } from './contexts/SnackbarProvider';
 export { default as VersionNavigator } from './components/VersionNavigator';
 export { default as CredentialForm } from './components/CredentialForm';
 export { default as SelectInputModal } from './components/SelectInputModal';
+export { VariablesProvider, useVariablesContext } from './contexts/VariablesProvider';
+export type { VariablesStore } from './contexts/VariablesProvider';
